# static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "InnoDD API",
    "version": "2.0.0",
    "description": "Complete InnoDD drug-discovery platform surface, generated from source (not the Postman collection).\n\n**Two surfaces:**\n- `/api/v1/*` — legacy internal pipeline modules (TxKG, Literature Mining, Drug Curation, Screening, Novelty Search) plus Auth/Users/Items scaffolding and job Orchestration. No X-Drp-Token needed.\n- `/v1/*` — the DRP frontend-facing contract (async job pattern: POST creates a job -> 202 {jobId}, then poll status/result). Needs `X-Drp-Token` (or `Authorization: Bearer`) from `/v1/auth/login`.\n\n**Known gaps / stubs (do not rely on for testing):**\n- `POST /api/v1/auth/login` and `/api/v1/auth/token` are unimplemented stubs (`pass`) in the legacy scaffold.\n- `/api/v1/users/*` imports a `UserRead` model that does not exist in `app/schemas/user.py` — this router likely fails to import and is NOT actually mounted.\n- `/api/v1/agentic/*` never mounts — `agent_routes.py` does not exist as a file; `_include()` swallows the ImportError.\n- Several module routers are double-mounted: Literature Mining, Drug Curation, Screening, and Novelty Search each exist BOTH under their versioned prefix (e.g. `/api/v1/literature-mining/...`) AND again with no prefix as a legacy alias (e.g. `/api/v1/search-keywords/`) — both paths are live simultaneously and hit the same handler.",
    "contact": { "name": "InnoDD API" }
  },
  "servers": [
...
  ],
  "components": {
    "securitySchemes": {
      "drpToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Drp-Token",
        "description": "Access token returned by POST /v1/auth/login or /v1/auth/refresh. The server also accepts a standard Authorization: Bearer header as a fallback."
      }
    },
    "schemas": {
...
      "SuccessResponse": { "type": "object", "properties": { "success": { "type": "boolean" } } },
      "UserCreate": { "type": "object", "properties": { "username": { "type": "string" }, "email": { "type": "string" }, "password": { "type": "string" } } },
...
      "UserUpdate": { "type": "object", "properties": { "username": { "type": "string" }, "email": { "type": "string" } } },
      "ItemCreate": { "type": "object", "properties": { "name": { "type": "string" }, "description": { "type": "string" } } },
      "ItemUpdate": { "type": "object", "properties": { "name": { "type": "string" }, "description": { "type": "string" } } },
      "Item": { "type": "object", "properties": { "id": { "type": "integer" }, "name": { "type": "string" }, "description": { "type": "string" } } },
...
      "ArticleDetail": {
        "type": "object",
        "properties": {
          "id": { "type": "string" }, "title": { "type": "string" }, "authors": { "type": "array", "items": { "type": "string" } },
          "year": { "type": "integer", "nullable": true }, "abstract": { "type": "string" }, "keywords": { "type": "array", "items": { "type": "string" } }, "pmcLink": { "type": "string", "nullable": true }
        }
      },
      "ExternalLink": { "type": "object", "properties": { "url": { "type": "string" }, "provider": { "type": "string" } } },
      "ChatRequest": { "type": "object", "required": ["message"], "properties": { "message": { "type": "string" } } },
      "ChatMessage": {
        "type": "object",
        "properties": { "role": { "type": "string" }, "content": { "type": "string" }, "citations": { "type": "array", "items": { "type": "string" } } }
      },
...
      "ProjectItem": {
        "type": "object",
        "properties": {
          "id": { "type": "string" }, "resultType": { "type": "string" }, "sessionId": { "type": "string", "nullable": true },
          "resultId": { "type": "string", "nullable": true }, "payload": { "type": "object" }, "createdAt": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/api/v1/users/": {
      "post": {
        "tags": ["Users (legacy scaffold)"], "summary": "Create user (router likely fails to import - UserRead missing)",
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserCreate" } } } },
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/api/v1/users/{user_id}": {
      "get": { "tags": ["Users (legacy scaffold)"], "summary": "Read user", "parameters": [{ "name": "user_id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "OK" } } },
      "put": {
        "tags": ["Users (legacy scaffold)"], "summary": "Update user",
        "parameters": [{ "name": "user_id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserUpdate" } } } },
        "responses": { "200": { "description": "OK" } }
      },
      "delete": { "tags": ["Users (legacy scaffold)"], "summary": "Delete user", "parameters": [{ "name": "user_id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "OK" } } }
    },
    "/api/v1/items/": {
      "post": {
        "tags": ["Items (legacy scaffold)"], "summary": "Create item",
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ItemCreate" } } } },
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Item" } } } } }
      }
    },
    "/api/v1/items/{item_id}": {
      "get": { "tags": ["Items (legacy scaffold)"], "summary": "Read item", "parameters": [{ "name": "item_id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Item" } } } } } },
      "put": {
        "tags": ["Items (legacy scaffold)"], "summary": "Update item",
        "parameters": [{ "name": "item_id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ItemUpdate" } } } },
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Item" } } } } }
      },
      "delete": { "tags": ["Items (legacy scaffold)"], "summary": "Delete item", "parameters": [{ "name": "item_id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "OK" } } }
    },
...
    "/v1/articles/{articleId}": {
      "get": {
        "tags": ["DRP Articles"], "summary": "Article detail", "security": [{ "drpToken": [] }],
        "parameters": [{ "name": "articleId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ArticleDetail" } } } } }
      }
    },
    "/v1/articles/{articleId}/pmc-link": {
      "get": {
        "tags": ["DRP Articles"], "summary": "Article PMC link", "security": [{ "drpToken": [] }],
        "parameters": [{ "name": "articleId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ExternalLink" } } } } }
      }
    },
    "/v1/articles/{articleId}/save": {
      "post": {
        "tags": ["DRP Articles"], "summary": "Save article to a project", "security": [{ "drpToken": [] }],
        "parameters": [
          { "name": "articleId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "projectId", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } } }
      }
    },
    "/v1/articles/{articleId}/chat": {
      "post": {
        "tags": ["DRP Articles"], "summary": "Chat with article (LLM or extractive fallback)", "security": [{ "drpToken": [] }],
        "parameters": [{ "name": "articleId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ChatRequest" } } } },
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ChatMessage" } } } } }
      }
    },
    "/v1/articles/{articleId}/chat/history": {
      "get": {
        "tags": ["DRP Articles"], "summary": "Article chat history", "security": [{ "drpToken": [] }],
        "parameters": [{ "name": "articleId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/ChatMessage" } } } } } }
      }
    },
...
    "/v1/projects/{projectId}/items": {
      "get": {
        "tags": ["DRP Projects"], "summary": "List project items", "security": [{ "drpToken": [] }],
        "parameters": [{ "name": "projectId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/ProjectItem" } } } } } }
      }
    },
...
  }
}